requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
python-docx>=0.8.11
PyMuPDF>=1.21.1
Pillow>=9.5.0
requests>=2.31.0
moviepy>=1.0.3
python-dotenv>=1.0.0
pandas==2.2.3
sqlalchemy==2.0.32